requirements_ultimate.txt
# ULTIMATE SPRITE PROCESSING PIPELINE - REQUIREMENTS (Windows Compatible)
# State-of-the-art AI-powered sprite processing with background removal,
# pose analysis, upscaling, and ComfyUI integration

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.0.0
scipy>=1.7.0

# Hugging Face AI Models
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
datasets>=2.10.0
tokenizers>=0.13.0

# Background Removal Models
rembg>=2.0.50
# backgroundremover>=0.2.6  # Sometimes problematic on Windows
segment-anything>=1.0

# Computer Vision & Image Processing
# mediapipe>=0.10.0  # COMMENTED OUT: Often fails on Windows
scikit-image>=0.20.0
imageio>=2.25.0
# imageio[ffmpeg]>=2.25.0  # COMMENTED OUT: Use separate ffmpeg installation

# Alternative to MediaPipe for Windows
ultralytics>=8.0.0  # YOLO models for pose detection
timm>=0.9.0         # Vision transformer models

# Advanced Upscaling (Windows compatible alternatives)
# realesrgan>=0.3.0      # COMMENTED OUT: Can be problematic on Windows
# basicsr>=1.4.2         # COMMENTED OUT: Heavy dependency
# facexlib>=0.3.0        # COMMENTED OUT: Often fails on Windows
# gfpgan>=1.3.8          # COMMENTED OUT: Heavy dependency

# ComfyUI Integration
websocket-client>=1.5.0
requests>=2.28.0
aiohttp>=3.8.0

# Deep Learning Utilities
einops>=0.6.0
safetensors>=0.3.0
# xformers>=0.0.20  # COMMENTED OUT: Optional, can be problematic on Windows

# Image Format Support
# webp>=0.1.5  # COMMENTED OUT: Often problematic
tifffile>=2023.0.0

# Progress & Logging
tqdm>=4.65.0
rich>=13.0.0
colorama>=0.4.6

# Data Processing
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.11.0

# Audio/Video Processing (for GIF optimization)
# moviepy>=1.0.3  # COMMENTED OUT: Heavy dependency, using imageio instead

# Development & Debugging
jupyter>=1.0.0
ipywidgets>=8.0.0

# Additional Windows-friendly packages
psutil>=5.9.0        # For system monitoring
pathlib>=1.0.1       # For path handling

# Optional Dependencies for Enhanced Features
# Uncomment ONLY if you can install them successfully on your system

# SUPIR Dependencies (for ultimate upscaling)
# omegaconf>=2.3.0
# lightning>=2.0.0

# Advanced Segmentation
# detectron2>=0.6.0  # Requires manual installation

# Professional Color Processing
# colour-science>=0.4.0

# GPU Memory Optimization
# bitsandbytes>=0.39.0  # For 8-bit inference

# Installation Notes for Windows:
# 1. Install PyTorch with CUDA support first:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# 2. Install ffmpeg separately:
#    Download from https://ffmpeg.org/download.html and add to PATH
#
# 3. For MediaPipe issues, we use YOLO models as alternative
#
# 4. Some models may require manual downloads:
#    - BiRefNet-HR: huggingface.co/ZhengPeng7/BiRefNet
#    - BRIA RMBG 2.0: huggingface.co/briaai/RMBG-2.0
#    - Real-ESRGAN models: Use built-in alternatives
#
# 5. For ComfyUI integration, ensure ComfyUI is installed separately
#
# 6. If any package fails to install, comment it out and the pipeline
#    will use fallback methods